+++ esp-hal/src/soc/esp32s3/mod.rs
@@ -23,6 +23,8 @@ crate::unstable_module! {
 }
 pub mod cpu_control;
 pub mod gpio;
+#[cfg(feature = "psram")]
+mod mmu;
 pub mod peripherals;
 
 /// The name of the chip ("esp32s3") as `&str`